+++ src/Console/SolanaStaking/Main.hs
@@ -169,7 +169,7 @@ argSpec =
         &= summary
             ( "solana-staking-csvs v"
                 <> showVersion version
-                <> ", Pavan Rikhi 2021"
+                <> ", Pavan Rikhi 2021-2024"
             )
         &= program "solana-staking-csvs"
         &= helpArg [name "h"]