+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## master
 
+
+
+## v0.1.3.0
+
 * Bump dependency versions(text).
 * Add `--aggregate` flag to group rewards by day.
 * Fetch rewards until no more rewards are returned instead of stopping when
   less than the 5-item page limit is returned.
 * Handle breaking API changes when rate limited.
+
 
 ## v0.1.2.0